# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reach Backend API",
    "version": "1.0.0",
    "description": "Auto-generated from src/routes/*.ts (Express + Zod). Re-run scripts/gen-openapi.py after changing routes."
  },
  "servers": [
    {
      "url": "https://chatapp-backend.onrender.com",
      "description": "Production (Render service: chatapp-backend -- edit if your actual URL differs)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clerk session token (requireAuth middleware)"
      }
    }
  },
  "paths": {
...
    "/api/internal/broadcast-moment-created": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Internal/broadcast moment created",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "momentId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "momentId"
                ]
              }
            }
          }
        }
      }
    },
    "/api/internal/broadcast-moment-deleted": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Internal/broadcast moment deleted",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "momentId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "momentId"
                ]
              }
            }
          }
        }
      }
    },
...
  }
}